requirements.txt
APScheduler==3.6.3
asgiref==3.7.2
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
idna==3.6
psycopg==3.1.14
python-telegram-bot==13.15
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tornado==6.1
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
